V2.11

+++ include/graphics.h
@@ -11,8 +11,8 @@ extern CHAR_INFO *consoleBuffer;
 
 void InitConsoleRenderer(unsigned int width, unsigned int height);
 void FreeConsoleRenderer();
-void SetCharOnPosition(unsigned int x, unsigned int y, char c, unsigned short color);
-void PrintStringOnPosition(char *s, unsigned int color, unsigned int x, unsigned int y);
-void SetCursorPosition(unsigned int x, unsigned int y);
+void SetCharOnPosition(int x, int y, char c, unsigned short color);
+void PrintStringOnPosition(char *s, unsigned short color, int x, int y);
+void SetCursorPosition(int x, int y);
 void WriteOutput();
 void ClearOutput();
